+++ CHANGELOG.md
@@ -19,6 +19,9 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 - New INI and Lua (R/W) `ACDropShip` property `HoverHeightModifier`. This allows for modification of the height at which an `ACDropShip` will hover when unloading cargo, or staying at a location.
 
 - New Lua event functions for `HDFirearm` - `OnFire(self)` that triggers when the gun fires, and `OnReload(self, hadMagazineBeforeReload)` that triggers when the gun is reloaded.
+
+- New `MOSRotating` INI and Lua (R/W) `MOSRotating` property `GibAtEndOfLifetime` that, when set to true, will make the `MOSRotating` gib if its age exceeds its lifetime, rather than deleting as it normally would.
+
 </details>
 
 <details><summary><b>Changed</b></summary>